requirements.txt
asn1crypto==1.2.0
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
cycler==0.10.0
Flask==1.1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
numpy==1.17.3
olefile==0.46
Pillow==6.2.1
pycparser==2.19
pyOpenSSL==19.1.0
pyparsing==2.4.5
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.13.0
torch==1.1.0
torchvision==0.3.0
tornado==6.0.3
urllib3==1.24.2
Werkzeug==0.16.0
win-inet-pton==1.1.0
wincertstore==0.2
